blog part 5

+++ app/Http/Controllers/Admin/AdminPostCategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Post;
 use App\Models\PostCategory;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -61,7 +62,13 @@ class AdminPostCategoryController extends Controller
     public function delete($id)
     {
         $row_data = PostCategory::where('id',$id)->first();
-        $row_data->delete();
+
+        $count = Post::where('post_category_id', $id)->count();
+        if($count==0) {
+            $row_data->delete();
+        } else {
+            return redirect()->back()->with('error', 'You can delete this post category, because there is one or more post under this.');
+        }
 
         return redirect()->back()->with('success', 'Data is deleted successfully.');
     }
